Restore main window position using SetWindowPlacement()
In e5797129448f48691d0da1107ddfcb717bbf85d7, the code used to create the
main window was updated to pass the saved window position directly to
CreateWindow().

The problem with that is that CreateWindow() accepts the position in
screen coordinates, while the window position is retrieved using
GetWindowPlacement() and is therefore saved in workspace coordinates.

Mixing these two coordinate systems meant that if the taskbar was in a
different position, the window would move each time the application
started. For example, if the taskbar was shown on the left of the
screen, the window would move to the left each time the application was
started, since workspace coordinates treated as screen coordinates would
be further to the left.

To fix this, the window position is now restored using
SetWindowPlacement().

Additionally, there is no longer an explicit check to see whether the
window is visible, since SetWindowPlacement() will adjust the
coordinates if they would result in the window being completely off
screen.