Fix issue with building "Bookmarks" menu
When building the "Bookmarks" menu, the items under the "Bookmarks Menu"
folder are inserted first, followed by the items in the "Other
Bookmarks" folder. A nextMenuId field is used to keep track of the next
available menu item ID.

Previously, that field wasn't being initialized, which meant that if
there were no "Bookmarks Menu" items, the field would never be set. When
inserting the "Other Bookmarks" items, the uninitialized field would be
used. Depending on the exact value of the field, no items might be added
(if the value was above the assigned range), or items might be added,
but not respond correctly when selected.

To fix this, the nextMenuId field is now initialized when building the
menu.