Don't open item on checkbox click
When the checkboxes option is enabled, a checkbox will be shown next to
each item. Clicking the checkbox for an item will allow the item to be
selected/deselected. When the single-click to open an item option is
enabled, items will be opened with a single click, rather than the
standard double click.

Previously, when both of those options were enabled at once, clicking on
a checkbox would result in the item being opened. Now, the click will be
ignored if it occurs over an item's state icon. When checkboxes are
enabled, the state icon will be used to represent the checkbox. This has
the effect of ignoring the click if it occurs over the checkbox.

This fixes #499.