Move group list into DirectoryState
Previously, the list of groups was managed by ShellBrowserImpl. However,
that meant that when a navigation occurred, the list wasn't being reset.
Because of that, the group count would increase each time a folder was
refreshed (as detailed in #481).

To fix this, the group list is now contained within DirectoryState and
will therefore be cleared whenever the current directory changes.

The issue has been present since
408b094a0d5baef09540f5221c37c566860a0072, as prior to that commit, the
groups would be cleared every time items were grouped. After that
commit, the groups were only cleared when group view was enabled.

The fix here is better than the original situation, since the groups are
only relevant within a folder and should be cleared whenever a
navigation occurs.

This fixes #481.