Ensure address bar is hidden when toggled off
To facilitate tab navigation, GetNextDlgTabItem() is used to find the
next window that should gain focus. That function will search for a
window that's visible, not disabled, and has the WS_TABSTOP style.

When the address bar is toggled off while the application is running,
RB_SHOWBAND is used to hide the associated band within the rebar. That
message will also cause the child window to be hidden, therefore
preventing it from participating in tab navigation.

When the address bar is toggled off on startup, however, the band was
hidden by setting the RBBS_HIDDEN style. Whilst that had the desired
effect (the band wouldn't be displayed), the child window was left as
visible, but with 0 size. That then meant that the window would
participate in tab navigation, even though it wasn't actually visible to
the user.

To prevent that, RB_SHOWBAND is now used to hide a band on startup, when
necessary. That ensures the window will be hidden.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?t=3461.