Fix languages specified in language DLL version headers
Each language DLL represents a particular language, with each language
identified by a language code that includes both the primary language ID
and the sublanguage ID.

In 18bb36fed2e10bb1ba9f01845aacae4cbce5caa8, the code used to load
language DLLs was updated, to use a fixed mapping between language codes
and DLL names. That mapping was based on the language code specified in
the main resource file for each DLL (i.e. Explorer++.rc).

Unfortunately, that mapping didn't take into account the fact that the
language is also specified in the version block for each DLL. It's that
language that is then used to verify that the DLL represents the
appropriate language and is also what's used to represent the selected
language.

For most of the languages, the language specified in the version block
didn't match the language specified in the main resource file. That
meant that, when loading a language, using a previously saved language
setting, the language code wouldn't appear in the fixed mapping and the
load would fail. The application would then revert back to English, as
reported in
https://explorerplusplus.com/forum/viewtopic.php?p=7512#p7512.

To fix that, there have been two specific changes that have been made:

1. The language codes specified in the version blocks have been updated
   to match the codes in the main resource files.
2. If a saved language code matches one of the codes that was updated,
   it will be automatically updated to the newer code. That ensures that
   if a custom language was saved using a previous version of the
   application, that setting will be honoured.