Fix issue that affects icons with overlays
When SHGetFileInfo() is used to retrieve the icon for an item, the
SHGFI_OVERLAYINDEX flag is specified, to ensure that the index of the
overlay icon is also returned.

Previously, the SHFILEINFO::iIcon value was used directly as the icon
index, which would work ok if there was no overlay. If there was an
overlay, however, using the iIcon value as an index wouldn't work, since
the upper 8 bits of iIcon would represent the overlay index. That then
meant that if a folder had an overlay icon, the icon for the folder
itself would fail to show up in a variety of places (e.g. in the tab
control, the manage bookmarks dialog, various menus).

To fix that, the icon index and overlay index are now extracted and
returned as separate values when fetching an icon. That makes it easier
to simply use the icon index, without having to first separate it from
the overlay index.

Additionally, some further CHECK's have been added.