Use concurrencpp to set timers in ShellTreeView class
Previously, when a timer needed to be set in the ShellTreeView class,
SetTimer() was used. However, setting a timer in that way has the
downside that the timer can be invoked even after KillTimer() has been
called (also see the more detailed explanation in
61a52d67965deff176149d0567078972237daa77).

That then meant that the application could crash when dragging an item
over the treeview, as described in #507.

To fix that, concurrencpp is now used to set timers in the ShellTreeView
class.

This fixes #507.