Format config.xml before save
Previously, the XML document was manually formatted. However, that
formatting was partial (since newer code didn't apply any formatting at
all) and tedious and error-prone to apply (since it required keeping
track of indentation levels).

The last of the manual formatting code was removed in
78d502becd71fcee5f7e2f1e0f75268873ee3828.

The solution here formats the document in an automated way, so it's
simpler and should work better.

Formatting the document isn't necessary, but is done purely as a
convenience for users.