Add ability to open a set of folders on startup
Along with the ability to open the default folder and open the previous
tabs, there is now an option to open a set of folders on startup. When
selected, the user can enter a set of paths that will be opened each
time the application starts.

Because there wasn't enough room for the UI to allow startup paths to be
added, the startup options have moved to their own page in the options
dialog. Previously, the startup options were displayed in the "General"
page, whereas they're now displayed in a "On Startup" page.

When the "Open a set of folders" option is selected on that page, paths
can be added to the associated list. The list always contains a blank
entry at the end, making it easy to add a path by double-clicking the
last item (which will put it into edit mode) and typing the path. Paths
can also be deleted by either erasing the text, or selecting the item
and pressing delete.