Use concurrencpp to perform work in ShellBrowserImpl directory change handling
When certain directory change notifications are received, there's work
that needs to be scheduled, but not run immediately. For example, if the
current folder is deleted, a navigation to the parent needs to occur.
That navigation can't, however, occur immediately, since
ShellBrowserImpl is in the middle of processing the change
notifications. The work needs to be deferred and handled at a later
point.

Previously, that was done by using some custom task code in
ShellBrowserImpl. Now, the work is performed using concurrencpp. This
has the advantage of using the same sort of style as other, recent,
asynchronous code. It also means that ShellBrowserImpl doesn't need to
have a custom task implementation.

One other improvement from this change is that there is some work that
was previously being done on the main thread that is now done on a
background thread.

For example, when navigating up to a parent item, it's first necessary
to find the nearest parent that exists. Previously, that was done on the
main thread. However, since finding the nearest parent will involve
accessing the disk, it makes more sense to do that from a background
thread.

A similar thing is also now done when the current directory is renamed.
The simple pidl is upgraded to a full pidl on a background thread.

Finally, the thread pool size in the ComStaThreadPoolExecutor class is
now set to std::thread::hardware_concurrency(), with a minimum size of
5. The numbers here are ultimately kind of arbitrary. The work here is
IO bound and blocking, so the ideal number of threads can be reliant on
details of the user's system and exactly how they use the application.
The number of threads can always be tweaked later on if it's
problematic.