Use WeakPtrFactory and WeakPtr in ShellBrowserImpl
Previously, weak references were generated by storing the
ShellBrowserImpl instance using a std::shared_ptr. However, it makes
more sense to store the instance using a std::unique_ptr, since the
lifetime isn't shared, and then pass out weak references using
WeakPtrFactory and WeakPtr.

This also has the advantage that when using the WeakPtr, checks will be
made to ensure that the access is occurring on the original thread.