Block nested tasks in UIThreadExecutor
Previously, when running tasks, UIThreadExecutor would implicitly allow
one task to run in the middle of another task. That behaviour, however,
can cause subtle reentrancy issues, since the nested task could invoke
functions in a reentrant way.

An example of the issues that can cause is described in #504. Although
the direct cause of that issue was slightly different, since it involved
the ShellBrowserImpl task queueing system (which was removed in
a0e53d80785727223d1ebd9f2b5bd02c47696fb1), the root cause was the fact
that tasks could be invoked in a nested way.

Therefore, UIThreadExecutor now blocks nested tasks. That means that if
one task is running, the next task will have to wait until the first
completes.

This also fixes an issue where tasks could be run out of order. If, for
example, there were two tasks queued and the first task entered a nested
loop, a third task could run in the middle of the first task, while the
second task would only run once the first task had completed.

That behaviour was due to the fact that UIThreadExecutor::OnTaskQueued()
would copy the task queue and run items from the copy. It effectively
meant that tasks would be batched and tasks from a single batch would be
ordered, but tasks between batches wouldn't be ordered.

This likely isn't a complete solution, since:

- There are still possible reentrancy issues when a nested message loop
  is entered. Since the nested message loop will pump messages, a task
  can run within the nested message loop, that is, in the middle of an
  existing block of code.
  For example, if a user-initiated navigation results in a dialog being
  shown, a task can run while the dialog is being shown and may initiate
  another navigation in the same tab.
- In some cases, however, it can be desirable to have tasks run within a
  nested message loop. For example, it's likely reasonable to have tasks
  run while a menu is being shown.

So, there will probably need to be more work done here.

This fixes #504.