Handle SHCNE_UPDATEITEM for current directory
The SHCNE_UPDATEITEM change notification is sent when an item is updated
in some way. For a directory, that can be when the icon changes, for
example, or the name changes (if the directory is virtual).

By handling the notification for the directory itself, it means that if
the icon or the name of the directory changes, those changes will now be
reflected throughout the UI. For example, the address bar will show the
updated name and icon.