Add initial entry on first navigation
In order to allow asynchronous navigation, there are a few changes that
have to be made first. One of those changes is to add a history entry as
soon as the first navigation starts.

Although that doesn't have any practical effect at the moment, since
navigation is synchronous, if navigation becomes asynchronous, it
ensures that there will always be a history entry, even if nothing has
been committed yet. That's important, since other parts of the
application rely on the fact that there is always at least one history
entry.

Once a committed entry is added, it will replace the existing initial
entry.