Change folders as soon as initial navigation starts
Typically, ShellBrowserImpl only changes folders when a navigation
commits. However, to support asynchronous navigation, it's necessary to
change folders when the initial navigation starts. That's to ensure that
ShellBrowserImpl always represents a folder.

Therefore, when the initial navigation starts, ShellBrowserImpl will now
change folders immediately.

Note that this technically implements the change requested in #514,
since when the initial navigation fails, the tab now remains in that
folder, regardless.

However, that issue has been left opened, since the behaviour here
should be consistent. That is, Explorer++ should always either navigate
away from a folder that doesn't exist, or remain in the folder. With
this change, some of that behaviour is now inconsistent, since
Explorer++ will navigate away from a folder when it's dynamically
deleted, but won't navigate away if the initial navigation fails.

Finally, this commit also fixes an issue in which the state of the
back/forward buttons in the manage bookmarks dialog weren't being
correctly updated when navigating back/forward.