Add BackgroundThreadEnumeration feature flag
This feature can be enabled using:

explorer++.exe --enable-features BackgroundThreadEnumeration

When enabled, the enumeration that takes place in ShellBrowserImpl will
occur on a background thread, rather than the main thread. This can be
very beneficial in cases where the folder takes a while to load (e.g. a
folder on a slow network).