Ensure current entry is correct during replacement
In f299d47f4f9e6330616fc593eeb96a03af709dc3,
ShellNavigationController::OnNavigationCommitted() was updated to
retrieve the current entry at the start of the function. The set of
selected items is then copied from the current entry at the end of the
function.

However, since the current entry can change in the body of the function,
the current entry wasn't necessarily current at the end. That then meant
that when navigating to a history entry, the selection wouldn't properly
be restored.

To fix that, the current entry is now retrieved again once it's been
updated. A test to ensure the selection is properly set has also been
added.