Add NavigationManager class
This class is responsible for managing ongoing navigations and allowing
new navigations to be started.

The primary advantage to this is that it allows a variety of
navigation-related code to be tested, in a way that wasn't really
possible before. For example, there are now tests that check whether the
appropriate signals are emitted during a navigation.