Remove ExecutionMode from NavigationManager
Previously, this was used to indicate whether the enumeration should be
synchronous or asynchronous. However, it makes more sense for the caller
to implicitly specify this via the executors passed in.

Also, previously, even if ExecutionMode::Async was passed in, along with
two executors, the enumeration wouldn't necessary be asynchronous. For
example, if both executors were inline, then the enumeration would
immediately run.