Set cursor in listview while active navigation is in progress
If there are any active navigations (an active navigation being a
navigation that can commit, once it returns to the main thread), the
cursor will now be changed to the IDC_APPSTARTING cursor via the
WM_SETCURSOR handler.

Until the NavigationManager class was added, there was a call to
SetCursor() when starting a new navigation. However, that isn't a good
solution, since the cursor could be over a completely different window
at that point. Additionally, the cursor would be reset as soon as it
moved.

By setting the cursor in the WM_SETCURSOR handler, the cursor will
remain the same, even when moved.

This is important when navigation is asynchronous, since the navigation
can take an arbitrary amount of time to complete in the background,
while the UI remains active.

To support this, some functionality has been added to NavigationManager
to allow the caller to detect whether there are any active navigations
in progress.