Update when loading text is displayed in status bar
Previously, when starting a navigation, a message would be displayed in
the status bar, indicating the folder that was being loaded. While that
works when navigation is synchronous, it doesn't really work when
navigation is asynchronous. For example, if the user switched to another
tab, then switched back, the loading message would no longer be shown.

Therefore, the status bar code has been updated to show the loading
message whenever there is an active navigation in a tab. That then means
that the message will be shown even if the user switches away from the
tab and then switches back.