Fix issue with toolbar theme in dark mode
When dark mode is enabled, toolbars previously had the DarkMode::Toolbar
theme implicitly applied. However, it appears that theme has issues that
cause it to not draw correctly.

Those issues could be seen when changing the DPI. In that situation, the
toolbar would re-open the theme, via a call to OpenThemeData(). If dark
mode was enabled, OpenThemeData() would then load the DarkMode::Toolbar
theme, which would cause button dropdowns (specifically the
TP_SPLITBUTTONDROPDOWN parts) to be shown incorrectly. A visual
depiction of that can be seen in #519.

One thing that's notable is that this issue only showed up once the DPI
was changed, regardless of whether dark mode was initially enabled. As
far as I can tell, that's because at the point at which the toolbars
were created, dark mode hadn't yet been enabled for the top-level
window. So, the theme opened during toolbar creation was always Toolbar.
If the DPI was later changed, the top-level window would have dark mode
enabled, which then meant that the DarkMode::Toolbar theme would be
opened, rather than Toolbar.

Note that the reproduction steps described in #519 are somewhat
different to the situation described above. In the issue, the problem
occurred after closing the laptop lid and re-opening it.

It appears that the toolbar class re-opens the theme in two situations:

- When the DPI is changed.
- When WM_THEMECHANGED is processed.

Therefore, if closing and re-opening the laptop lid resulted in either
of those events occurring, that would explain why the dropdown buttons
abruptly changed.

Once the DarkMode::Toolbar theme was applied, it would also cause other
visual issues. For example, switching back to light mode would result in
the toolbars drawing incorrectly (there would be a mixture of light and
dark colors). That's because the DarkMode::Toolbar theme would continue
to be applied, since, as noted above, the toolbar class only re-opens
the theme in response to specific events. Disabling dark mode on the
window isn't one of those events.

To fix this, the following call is now made when theming a toolbar:

SetWindowTheme(hwnd, L"", nullptr);

This ultimately stops the DarkMode::Toolbar theme from being applied. It
appears that any app name (the second argument) works, provided it
doesn't match an actual theme name. Presumably that's because there's
fallback within the theme code. So, specifying an app name that doesn't
match an actual theme results in a fallback to the default theme (i.e.
Toolbar). The default theme can be used in both light mode and dark
mode, since the appropriate dark mode colors are applied in other ways
(e.g. by custom drawing).

This fixes #519.