Commit initial navigation if it's stopped
Typically, if a navigation is stopped, it will be cancelled once it
returns to the main thread, and the original folder will continue to be
shown. However, if the initial navigation is stopped, there won't be any
original folder, so it makes more sense to commit the navigation.
Therefore, if the initial navigation is stopped, it will now be
committed, rather than cancelled.

This also matches what happens when the initial navigation fails - it
will be committed, rather than failed.