Remove navigation completion event
This simplifies the navigation process in the success case. It also
means that the success case is closer to the failure/cancellation cases.
That is, once the committed event is broadcast, the navigation is
complete, in the same way that it's complete once the failed/cancelled
event is broadcast.

Additionally, NavigationManager now simply supplies the list of items in
the navigation committed event, rather than using a separate items
available event.