Start tab navigation after tab creation event has been triggered
Previously, the navigation was started before the tab creation event was
triggered. That's necessary because a tab always requires a current
directory. Previously, there was no current directory until the first
navigation was started.

Now, the current directory is set when the tab is created. The way it's
set (by passing it the pidl that will ultimately be used for the first
navigation) isn't ideal, since there's no direct connection between the
pidl passed to the ShellBrowserImpl constructor and the pidl used for
the first navigation.

Web browsers have the ability to use about:blank as a fallback when
there isn't any other URL, but there isn't an analogous filesystem path.
Potentially, a simple pidl could be created and used as a fallback, but
that could have unforeseen impacts on other functionality (e.g. how the
background context menu functions, what happens when shell operations
are invoked, what happens when the tab is refreshed, or when the path is
manually typed out, etc).

Using the pidl that will actually be used for the first navigation makes
more sense, given those concerns.

Overall, that now means that the first navigation is started after the
tab creation has been triggered. That's more in line with what would be
expected. The previous solution, in which the navigation was started
(and also completed, if synchronous) before the tab creation was
triggered was unusual.

Additionally, the tab selection event is now also triggered after the
tab creation event, rather than before.