Update how active tab check is performed in address bar
Previously, this was checked by calling TabContainer::IsTabSelected().
However, the disadvantage of that is that if there are multiple panes
within a browser window, the question of whether a tab is selected in
its container is different to the question of whether the tab is active
within its browser window. That is, there could be multiple selected
tabs, with only one of them being active.

Therefore, the active tab is now checked by consulting the browser.