Broadcast filter event through ShellBrowserEvents
When items are filtered/unfiltered, the displayed items will change.
That event is now broadcast through ShellBrowserEvents, rather than via
a custom message.

The signal that's used is the same as the one that's used when the
directory contents change. That's because both events result in the
displayed items changing.