Remove IDirectoryMonitor
Initially, this was being used for directory monitoring. However, given
that it only works for filesystem folders (due to the use of
ReadDirectoryChangesW), an alternative solution was put in place that
covered all shell folders (via use of SHChangeNotifyRegister). That
implementation has been the default for a while.

Additionally, the implementation of IDirectoryMonitor had various
multi-threaded issues and was generally poor quality.

Therefore, IDirectoryMonitor has now been removed.

Note that SHChangeNotifyRegister can't really be used on Windows PE,
since it requires Explorer to be running. Therefore, a different
implementation, built on top of ReadDirectoryChangesW, will be put in
place for use on Windows PE in a future commit.