Add FileSystemChangeWatcher class
This class watches for directory changes using ReadDirectoryChangesW()
(indirectly, via WIL). This is useful on Windows PE, where the default
watcher (ShellChangeWatcher) doesn't work, since
SHChangeNotifyRegister() requires Explorer to be running.

A --change-notify-mode command line option has been added to allow the
change implementation to be selected at runtime. On Windows PE, this
option will default to "filesystem", in any other case "shell". This
will ensure that directory change notifications work on Windows PE, for
at least filesystem folders.

This is a partial fix for #403. Partial since the treeview will also
need to be updated to make use of FileSystemChangeWatcher when
necessary.