Register for file change notifications in ShellTreeView
The ShellTreeView class displays a hierarchical list of folders.
However, certain types of container files (e.g. e.g. .7z, .cab, .rar,
.zip) act like folders as well and are also displayed.

Previously, file change notifications weren't observed at all. That
meant that, for example, a zip file would be added to the treeview
during the enumeration process, but wouldn't be updated at all if it was
renamed/removed. Similarly, if a new zip file was added, it wouldn't be
added to the treeview, unless the parent node was collapsed and
re-expanded (which would trigger a fresh enumeration).

To fix those sorts of issues, file changes are now observed by
ShellTreeView. Note that a file will only be added to the treeview if it
has the SFGAO_FOLDER attribute set.