Update how theme is applied to dialogs
Previously, all dialogs inherited from ThemedDialog. That class then
instantiated a ThemeWindowTracker instance on dialog creation. The
downside of that was that it meant that the ThemeManager instance had to
be passed to every dialog, along with every function that directly or
indirectly created a dialog.

Now, the dialog creation methods in Win32ResourceLoader will instantiate
the ThemeWindowTracker instance when necessary. That means that
ThemeManager only needs to be passed to one class (Win32ResourceLoader),
rather than many classes.

The other advantage of this change is that, when creating a dialog,
callers can now pass the dialog procedure in via a std::function, rather
than via a raw function pointer.