Fully draw tab control in dark mode
This ensures that the tab control is drawn in dark colors when dark mode
is enabled.

Additionally, the up-down control that appears when the control can be
scrolled is now also themed. Note that because the Window theme that's
set on the up-down control is only implemented in Windows 11, the
control will still appear in light colors on Windows 10 when dark mode
is enabled.

This fixes #508 and is part of the work for issue #115.