Move dark mode color constants into a separate class
This is an improvement on the previous situation, since the colors are
now only stored in a single class (previously, the colors were split
between DarkModeManager and ThemeManager). Additionally, it makes more
sense for DarkModeManager to manage dark mode from an OS standpoint,
rather than also managing some of the dark mode colors.