Update background color of checked toolbar buttons in dark mode
A checked button is now marked as hot when custom drawing (an idea
suggested in
https://github.com/derceg/explorerplusplus/issues/115#issuecomment-2274287659).

This is an improvement over the previous situation, since the default
background color for a checked toolbar item is too bright. Additionally,
checked item will now retain the same background color even if they're
actually hot, which matches the behaviour of the control under the
default theme.

This is part of the work for issue #115.