Correct call to SHGetFileInfo() in AddPathsToComboBoxEx()
Previously, the arguments being passed to SHGetFileInfo() were
incorrect.

Additionally, the icon for the current directory was being repeatedly
retrieved, rather than the icon for the item itself. That was causing
the same icon to be used throughout the entire dropdown.

That has now being fixed, so that the icon for each item matches the
associated file's icon.