Add BrowserCommandTargetManager class
Some commands (e.g. cut, copy, paste) are context-sensitive, in that
they act on the focused window. At the moment, these commands are all
handled in the Explorerplusplus class, which has a few downsides:

- It breaks encapsulation, since details that could potentially be
  hidden by a class (e.g. information about selected items) has to be
  provided to Explorerplusplus.
- It means that the Explorerplusplus class has to be aware of which
  windows can participate in specific actions. For example, it has to
  know that the copy command can be invoked for the listview/treeview,
  but not for the address bar.

The aim of the BrowserCommandTargetManager class, along with some
associated classes, is to allow the Explorerplusplus class to invoke
context-sensitive commands in a generic way, without having to
explicitly depend on particular views.

Not that BrowserCommandTargetManager and its associated classes aren't
used at the moment.