Observe focus changes via BrowserCommandTargetManager
One advantage of this is that if the target hasn't changed, no event
will be broadcast. For example, when re-activating a browser window, a
WM_SETFOCUS message will be sent. However, if the window that's
receiving focus was the same window that had focus before the browser
window was deactivated, there's no need to broadcast a target changed
event, since the target hasn't actually changed.