Update method used to detect whether a new folder can be created
The updated method better aligns with what the shell is doing to
determine whether a new folder can be created. Additionally, the updated
method correctly handles container files.

Previously, CanCreateInDirectory() would return true for all filesystem
items, which meant it would return true for container files. That's not
correct, however, as it's not possible to create a new folder within a
container file. The updated method correctly indicates that a new folder
can't be created within a container file.

It also better handles read-only directories. Note that whether a
directory is read-only is independent of the FILE_ATTRIBUTE_READONLY
attribute (which has no meaning for directories). A read-only directory
can be created, for example, by mounting a virtual hard disk as
read-only.

As with container files, the previous version of CanCreateInDirectory()
would return true for read-only directories if they were filesystem
items. That's not correct, as it's not possible to create new items in
read-only directories. The updated method handles this case correctly.