Fix issue with navigation target mode not being set
In fd1e729fd50d0b1c9a2fd0ad0150ffb7eee0c840, the Tab class was updated
to accept the lock state in its constructor. When the lock state is set,
the Tab class is then responsible for setting the navigation target mode
on the ShellBrowser instance.

That wasn't being done during construction, however. That meant that if
the tab was locked after construction, the navigation target mode would
be set correctly. If it was locked during construction, however, the
target mode wouldn't be set. That then meant that navigations in the tab
would proceed in that tab, rather than in a new tab.

To fix that, the target mode is now set during tab construction.

This fixes #541.