Don't update status bar text if menu help text is being shown
There can be situations where a menu is being shown and an event occurs
that would typically cause the status bar to update. For example, if a
context menu is being shown and the back button on a mouse is clicked,
the current tab can navigate back while the menu is being shown.

In those situations, the status bar doesn't need to be updated, since
it's displaying the menu help text. Additionally, attempting to update
the text will trigger a DCHECK failure, since some of the calls to
SB_SETTEXT will fail (due to the status bar parts being different).

Therefore, the status bar text will no longer be updated if menu help
text is being shown.