Handle command prompt commands in BrowserCommandController
This also fixes an issue in which a command prompt could be launched
even when the current directory was the top-level folder in a container
file. Within a nested folder in a container file, the SFGAO_FILESYSTEM
attribute won't be set. However, that attribute will be set for the
top-level folder.

Because the command prompt commands were previously disabled for any
non-filesystem folder, that meant that the commands would be disabled if
the current directory was a nested folder in a container file.

However, if the current directory was the top-level folder in a
container file, the command prompt commands would be enabled.

While container files can be navigated to within Explorer++, the
contents are effectively virtual, so it's not possible to start a
command prompt in a container file directory.

Therefore, the command prompt commands will now be disabled if the
current directory is the top level of a container file.