Handle split file command in BrowserCommandController
This also fixes an issue with how the split file menu item and split
file toolbar item were enabled/disabled. In both cases, there was a
check to determine whether a single item was selected. However, when
setting the state of the main menu item, it was also checked whether or
not the current folder was virtual.

Since this check is now performed by a single method, this inconsistency
has been fixed.

Additionally, the command will now only be enabled if the current
directory isn't the top-level of a container file.