Replace PidlBase::TakeOwnership() with a constructor
TakeOwnership() was awkward, in that it meant that the caller couldn't
declare and initialize a pidl in one step. Instead, the pidl had to be
declared and then TakeOwnership() had to be called separately.

Using a constructor means that a pidl can be declared and initialized in
a single step.

The new constructor takes a token type, to ensure that it isn't going to
be confused with the other constructors.