Update how shell context menu is managed
Previously, when creating a ShellContextMenu instance, a single delegate
could be provided. Because there was only a single delegate, that design
had the downside that each caller would have to add and handle menu
items on their own.

Certain items (e.g. the "Open in new tab" item) should be shown whenever
a context menu is displayed (depending on exactly which item the context
menu is being shown for). Other items (e.g. the built-in "Open" item)
should be always handled internally.

Both of those things were more difficult to handle with the previous
design. Because only a single delegate could be provided, each caller
would have to update the menu, leading to duplicated code.

Now, multiple delegates can be provided. That has the advantage that the
code to add/handle particular items can be centralized in an individual
delegate. Multiple callers can then use that delegate, without having to
do the work themselves.

This commit also improves how item IDs are handled for the shell context
menu. When the menu is retrieved, the shell items will be assigned an ID
range. Custom items should then use IDs outside of that range.

Previously, there was no explicit handling for that. That meant that,
for example, the custom items on the background menu would simply be
assumed not to clash. In other cases, menu IDs were assigned just
outside the shell item range, but there was nothing enforcing that.

Now, delegates aren't given access to the actual menu. Instead, they can
only modify it through a builder interface that ShellContextMenu
provides. When the delegate adds an item, the builder will remap the ID
internally. Then, when the delegate needs to do something with a menu
item (e.g. respond to a selection), the ShellContextMenu can provide the
original ID.

This design ensures that delegates only interact with the IDs they use,
even though the IDs will be remapped for use on the menu. That also
means that delegates don't have to do anything special to support ID
remapping.