Update how simulated clipboard works
Previously, a SimulatedClipboard class was added. That class provided
some methods to read and write data to a simulated clipboard.

However, rather than basing the simulated class around the interface
used to read and write data to the clipboard, it makes more sense to
base such a class on the underlying clipboard store. There's a few
advantages to doing that:

- It's simpler. The simulated store only has to implement a few methods,
  rather than trying to implement a variety of methods to read/write
  data in different formats.
- The store can be passed to the existing Clipboard class, allowing the
  full functionality from that class to be used.
- Since there's only going to be a single store, a single instance can
  be created and passed around throughout the application.

The simulated store is now also used in a few unit tests. This helped
pick up a bug in the BookmarkContextMenuSingleBookmarkTest::Paste test,
where the test was implicitly relying on the fact that earlier tests
copied bookmark data to the clipboard. Switching to the simulated store
helped reveal that, since the simulated store is always empty initially.